fix : fix wrong placing click event handler on tab works section
the previous code version, attach click event on the tab wrapper the detect the e.target id then display the tab content based on the id detected.
but it generated error if user click on the child of tab-title, so the id's not detected.
i change to directly attach to the child (the tab title)